fixed hammer.js and delete objects

+++ src/app/client/pages/canvas/draw/components/board.vue
@@ -33,7 +33,6 @@
             <i class="iconfont" @click="changeZoom()">&#xe663;</i>
           </li>
         </ul>
-
       </div>
       <div class="tools">
         <ul>
@@ -43,6 +42,7 @@
              :class="{'selected': plugin.active}"
              class="plugin-tools-item"
              :title="plugin.title">
+
             <i class="iconfont" :class="{'disabled': !plugin.useInFollowing && notPresenter}" v-html="plugin.icon"></i>
             <template v-if="plugin.hasAction">
               <component
@@ -54,7 +54,6 @@
               :is="key + '-action'" >
               </component>
             </template>
-
           </li>
           <!-- <li><i class="icon ion-md-brush"></i></li> -->
         </ul>
@@ -483,7 +482,9 @@ export default {
       !opid && this.socket.emit('sync', 'undo', item, this.board._id, this.board._id)
     },
     deleteSelected() {
-      this.drawer.deleteSelected()
+      if (this.canDelete) {
+        this.drawer.deleteSelected()
+      }
     },
     choose(chooseKey, hiddenAction) {
       if (!this.plugins[chooseKey].useInFollowing && this.notPresenter) {
